test : ChatLogService 테스트 생성
